GameNews_bot/requirements.txt
anthropic>=0.40.0
google-genai>=1.0.0
ddgs>=0.5.0
duckduckgo-search>=6.2.0
requests>=2.31.0
